Update README.md
Summarize the project and what problem it was solving.  How to Get an Instance Attribute/ How to Update an Instance Attribute. When you change the instance variable’s values of one object, the changes will not be reflected in the remaining objects because every object maintains a separate copy of the instance variable. Code is read much more often than is written that is why is so important to stick to conventions that will help us or other people to understand more easily the available code. 
What did you do particularly well? Well I found that this particular tool helps find bad practice such as -Arguments inside the parenthesis are not aligned vertically, or using hanging indentation. Also is detects  a wrong indentation. Good practices highlights arguments inside the parenthesis are aligned vertically.
Where could you enhance your code? How would these improvements make your code more efficient, secure, and so on? So I ve learned that I could possibly improve on always surrounding my operators with a single space on either side such as : Assignment (=)
Augmented assignment (+=, -=, etc)
Comparison operators (<, >, <=, >=, ==, !=)
Logical operators (and, or, not)
Membership operators (in, not in)
Identity operators (is, not is)
Which pieces of the code did you find most challenging to write, and how did you overcome this? What tools or resources are you adding to your support network? So previous projects that I have done my problems where with import error and key error. 
What skills from this project will be particularly transferable to other projects or course work?If I had this type of tool or would have know about it maybe it could have helped me with so many of my errors.
How did you make this program maintainable, readable, and adaptable?I remember my professor saying less is more because It makes sense to remember that the less code you have, the less there is to maintain. Also good error checking keeps your program from exploding and makes it faster to debug. When you take a look at a procedure where all the error checking is in order, you don’t have to spend time seeking the mistake. Lastly coming code you should avoid commenting in ways that no one can understand (not even you) will eventually turn against you. I read that commenting your code is of utmost importance because if you write a procedure and fail to comment it, the lack of comments will cost you time when you decide to rework it.